% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fpkm_files_from_excel_data_functions.R
\name{add_new_data_to_fungiexpresz}
\alias{add_new_data_to_fungiexpresz}
\title{Add new data to FungiExpresZ.}
\usage{
add_new_data_to_fungiexpresz(new_data_matrix = data_matrix,
  new_sample_info = sample_info,
  new_data_matrix_rds_path = "~/Documents/Projects/8_FungiExpresZ_iMac_version/fungiexpresz_r_package/FungiExpresZ/inst/app/expression_mats_rds_files_new/")
}
\arguments{
\item{new_data_matrix}{a tibble of gene expression generated using the function [FungiExpresZ::prepare_fungiexpresz_fpkm_matrix_rds_files_from_stringTie_output()].}

\item{new_sample_info}{a tibble of SRA annotations associated with `new_data_matrix`. This object can be generated using the function [FungiExpresZ::prepare_annotations_for_fungiexpres_data_matrix()].}

\item{existing_sample_info_rds}{a character string denoting path to existing `sample_info`.}

\item{existing_reference_annotation_to_expr_map_rds}{}
}
\description{
Add new data to FungiExpresZ.
}
