% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{app server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{session input}

\item{output}{session output}

\item{session}{session}
}
\description{
app server
}
\keyword{internal}
