% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{badge_custom_url}
\alias{badge_custom_url}
\title{get custom badge url}
\usage{
badge_custom_url(x = "current", y = NULL, col = "red", add_github_logo = FALSE)
}
\arguments{
\item{x}{text to show on left panel of badge}

\item{y}{text to show on right panel of badge}

\item{col}{badge color}

\item{add_github_logo}{TRUE if github logo to be shown on badge}
}
\value{
url
}
\description{
get custom badge url
}
\keyword{internal}
