% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{badge_devel_url}
\alias{badge_devel_url}
\title{get github repo version devel bedge}
\usage{
badge_devel_url(pkg = "cparsania/fungiexpresz", col = "green",
  add_github_logo = FALSE)
}
\arguments{
\item{pkg}{github repo url}

\item{col}{bedge color}

\item{add_github_logo}{TRUE if github logo to be shown on badge}
}
\value{
url
}
\description{
get github repo version devel bedge
}
\keyword{internal}
