% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{badge_github_release_url}
\alias{badge_github_release_url}
\title{create a custom badge url for github release}
\usage{
badge_github_release_url(
  url = "cparsania/fungiexpresz",
  col = "red",
  logo = "github",
  logo_color = NULL
)
}
\arguments{
\item{url}{github repo url, default : cparsania/fungiexpresz}

\item{col}{badge color, default : red}

\item{logo}{badge logo , default : github. Few choices are apple, windows, git, docker etc.}

\item{logo_color}{logo color , default : default logo color}
}
\value{
badge url
}
\description{
create a custom badge url for github release
}
\keyword{internal}
