% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{check_var_types}
\alias{check_var_types}
\title{validate vars}
\usage{
check_var_types(x, req_data_type)
}
\arguments{
\item{x}{: tibble of vars to check}

\item{req_data_type}{: req data type of each col in x. Must have same lenght to ncol(x)}
}
\value{
internal
}
\description{
validate vars
}
\keyword{internal}
