% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{cross_ref_go_db}
\alias{cross_ref_go_db}
\title{Cross reference GO_ID to GID mapping from GO.db data}
\usage{
cross_ref_go_db(term_to_gene)
}
\arguments{
\item{term_to_gene}{a tbl with two columns having many to many mapping between GO_ID and GID : Column 1 GO_ID and column 2 GID}
}
\value{
a tbl with many to many mapping.
}
\description{
given a GO_ID GID mapping, this function cross reference GO_ID with GO.db. For each GO_ID, it map all GID for all offspring GO_ID
}
\keyword{internal}
