% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{decorate_ggplot}
\alias{decorate_ggplot}
\title{Decorate GG plot}
\usage{
decorate_ggplot(
  gplot,
  x_labs_size,
  y_labs_size,
  x_tick_size,
  y_tick_size,
  x_tick_angle = 90,
  title_labs_size,
  plot_title = "",
  axis_x_title = "",
  axis_y_title = "",
  color_legend_title = "",
  fill_legend_title = "",
  legend_text_size = 15,
  legened_title_size = 15,
  legend_key_height = 10,
  strip_text_size = 15,
  legend_key_width = 10,
  legend_spacing_x = 0.2,
  legend_spacing_y = 0.2,
  guide_legend = TRUE,
  ...
)
}
\arguments{
\item{gplot}{an object of class ggplot2}

\item{x_labs_size}{internal}

\item{y_labs_size}{internal}

\item{x_tick_size}{internal}

\item{y_tick_size}{internal}

\item{x_tick_angle}{internal}

\item{title_labs_size}{internal}

\item{plot_title}{internal}

\item{axis_x_title}{internal}

\item{axis_y_title}{internal}

\item{color_legend_title}{internal}

\item{fill_legend_title}{internal}

\item{legend_text_size}{internal}

\item{legened_title_size}{internal}

\item{legend_key_height}{internal}

\item{strip_text_size}{internal}

\item{legend_key_width}{internal}

\item{legend_spacing_x}{internal}

\item{legend_spacing_y}{internal}

\item{guide_legend}{internal}

\item{...}{other parameters passed to \code{ggplot2::theme}}
}
\value{
internal
}
\description{
Decorate GG plot
}
\keyword{internal}
