% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fpkm_files_from_excel_data_functions.R
\name{.filter_samples_by_mapping_rate}
\alias{.filter_samples_by_mapping_rate}
\title{Filter FPKM matrix by mapping rate}
\usage{
.filter_samples_by_mapping_rate(x, mapping_threshold = 0.7)
}
\arguments{
\item{x}{a character string denoting a file to mapping rate.}

\item{mapping_threshold}{threshold value, default 0.7.}
}
\value{
a tibble.
}
\description{
Filter FPKM matrix by mapping rate
}
\keyword{internal}
