% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fpkm_files_from_excel_data_functions.R
\name{.get_annotations_from_fungal_sra_run_info_table}
\alias{.get_annotations_from_fungal_sra_run_info_table}
\title{Prepare SRA sample annotations from fungal_sra_run_info_table}
\usage{
.get_annotations_from_fungal_sra_run_info_table(sra_id,
  fungal_sra_run_info_object_path)
}
\arguments{
\item{sra_id}{a character vector denoting valid SRA id.}

\item{fungal_sra_run_info_object_path}{a character string denoting path to the `fungal_sra_run_info_table.rds` object}
}
\value{
a tibble.
}
\description{
Prepare SRA sample annotations from fungal_sra_run_info_table
}
\details{
this is an internal function. It is called from the [FungiExpresZ::prepare_annotations_for_fungiexpres_data_mtrix()].
}
\keyword{internal}
