% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fpkm_files_from_excel_data_functions.R
\name{.get_annotations_from_sradb}
\alias{.get_annotations_from_sradb}
\title{Prepare SRA sample annotations from  SRAdb}
\usage{
.get_annotations_from_sradb(sra_id, path_sra_db_sqlite_file)
}
\arguments{
\item{sra_id}{a character vector denoting valid SRA id.}

\item{path_sra_db_sqlite_file}{a character string denoting path to the `SRAmetadb.sqlite` object.}
}
\value{
a tibble.
}
\description{
It use three tables - `sra`, `experiment` and `sample` to get necessary information for an SRA id.
}
\details{
this is an internal function. It is called from the [FungiExpresZ::prepare_annotations_for_fungiexpres_data_mtrix()].
}
\keyword{internal}
