% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{generate_random_strings}
\alias{generate_random_strings}
\title{Generates n random string of length 10}
\usage{
generate_random_strings(n = 5000)
}
\arguments{
\item{n}{An integer denoting number of unique strings to be generated.}
}
\value{
A character vector of length \code{n} containing unique string
}
\description{
Generates n random string of length 10
}
\keyword{internal}
