% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_ca_gene_names_mapping}
\alias{get_ca_gene_names_mapping}
\title{Map Candida albicans genes between orf_XXXXX and C1_XXXXX}
\usage{
get_ca_gene_names_mapping()
}
\description{
Get tibble of Candida albicans gene names map (orfXXX to C1_XXXX)
}
\keyword{internal}
