% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_density}
\alias{get_density}
\title{get_density}
\usage{
get_density(x, y, n = 200)
}
\arguments{
\item{x}{internal}

\item{y}{internal}

\item{n}{internal}
}
\value{
internal
}
\description{
calculate 2D density
}
\keyword{internal}
