% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_expression_mats_dir_path}
\alias{get_expression_mats_dir_path}
\title{Get absolute path for the expression matrix .rds files. 
It returns the character string indicating to absolute directory path of expression matrix .rds files.}
\usage{
get_expression_mats_dir_path()
}
\value{
A character vector of a string.
}
\description{
Get absolute path for the expression matrix .rds files. 
It returns the character string indicating to absolute directory path of expression matrix .rds files.
}
\keyword{internal}
