% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_gg_colors}
\alias{get_gg_colors}
\title{Get character vector of  ggplot colors in default order}
\usage{
get_gg_colors(x)
}
\arguments{
\item{x}{internal}
}
\value{
internal
}
\description{
Given a character or numeric vactor \code{x} , the function returns the named vector of default ggplot colors of length \code{length(x)}. The names will be elements of \code{x} and values will be unique color associated to each element of \code{x}.
}
\examples{
\dontrun{
x <- letters[1:5]
gg_col = get_gg_colors(x)

## visualize colors 
pie(seq_len(length(x)) / seq_len(length(x)) ,labels = x , col=gg_col) 

y <- LETTERS[ 1:5]

visualize colors 
get_gg_colors(y)
pie(seq_len(length(y)) / seq_len(length(y)) ,labels = y , col=gg_col)
} 
}
\keyword{internal}
