% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_test_data}
\alias{get_test_data}
\title{get random sample of gene expression matrix}
\usage{
get_test_data(n_samp = 5)
}
\arguments{
\item{n_samp}{An integer denoting number of SRA samples to sample.}
}
\value{
internal
}
\description{
get random sample of gene expression matrix
}
\keyword{internal}
