% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fpkm_files_from_excel_data_functions.R
\name{get_updated_existing_annotations}
\alias{get_updated_existing_annotations}
\title{Get updated annotations objects (sample_info & ref_annot_to_expr_rds) of FungiExpreZ.}
\usage{
get_updated_existing_annotations(new_sample_info = sample_info)
}
\arguments{
\item{new_sample_info}{a tibble of SRA annotations associated with `new_data_matrix`. This object can be generated using the function [FungiExpresZ::prepare_annotations_for_fungiexpres_data_matrix()].}
}
\value{
a list of two -> updated_ref_annot_to_expr_rds, updated_sample_info.
}
\description{
Get updated annotations objects (sample_info & ref_annot_to_expr_rds) of FungiExpreZ.
}
\details{
This function returns a list of two objects - 
1) updated_ref_annot_to_expr_rds: This is the updated version of existing FungiExpresZ object -> `ref_annot_to_expr_rds`.
2) updated_sample_info: This is the updated version of existing FungiExpresZ object -> `sample_info`.
Both these objects will be added R/sysdata.rda.

Internally it uses FungiExpresZ:::sample_info and FungiExpresZ:::ref_annot_to_expr_rds to get the access to existing annotaions.
}
