% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{lm_eqn}
\alias{lm_eqn}
\title{lm_eqn}
\usage{
lm_eqn(df, x, y)
}
\arguments{
\item{df}{internal}

\item{x}{internal}

\item{y}{internal}
}
\value{
internal
}
\description{
regression eqn
}
\keyword{internal}
