% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_add_gene_groups.R
\name{add_gene_groups_ui}
\alias{add_gene_groups_ui}
\alias{add_gene_groups}
\title{add_gene_groups_ui and add_gene_groups}
\usage{
add_gene_groups_ui(id)

add_gene_groups(input, output, session, gene_names)
}
\arguments{
\item{id}{shiny id}

\item{input}{session input}

\item{output}{session output}

\item{session}{session}

\item{gene_names}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
