% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_add_sample_group.R
\name{add_sample_group_ui}
\alias{add_sample_group_ui}
\alias{add_sample_group}
\title{add_sample_group_ui and add_sample_group}
\usage{
add_sample_group_ui(id)

add_sample_group(input, output, session, sample_names, bioproject_groups)
}
\arguments{
\item{id}{shiny id}

\item{input}{session input}

\item{output}{session output}

\item{session}{session}

\item{sample_names}{internal}

\item{bioproject_groups}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
