% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cluster_wise_sample_information.R
\name{cluster_wise_sample_information_ui}
\alias{cluster_wise_sample_information_ui}
\alias{cluster_wise_sample_information_server}
\title{cluster_wise_sample_information_ui and cluster_wise_sample_information_server}
\usage{
cluster_wise_sample_information_ui(id)

cluster_wise_sample_information_server(
  input,
  output,
  session,
  parent_id,
  cluster_wise_sample_information
)
}
\arguments{
\item{id}{shiny id}

\item{input}{session input}

\item{output}{session output}

\item{session}{session}

\item{parent_id}{internal}

\item{cluster_wise_sample_information}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
