% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_active_groups.R
\name{display_active_groups_ui}
\alias{display_active_groups_ui}
\alias{display_active_groups_server}
\title{display_active_groups_ui and display_active_groups_server}
\usage{
display_active_groups_ui(id)

display_active_groups_server(input, output, session, group_data)
}
\arguments{
\item{id}{shiny id}

\item{input}{session input}

\item{output}{session output}

\item{session}{session}

\item{group_data}{}
}
\description{
A shiny Module.
}
\keyword{internal}
