% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_display_sra_sample_info.R
\name{display_sra_sample_info_ui}
\alias{display_sra_sample_info_ui}
\alias{display_sra_sample_info_server}
\title{display_sra_sample_info_ui and display_sra_sample_info_server}
\usage{
display_sra_sample_info_ui(id)

display_sra_sample_info_server(input, output, session,
  user_selected_sra_sample_info_tibble)
}
\arguments{
\item{id}{shiny id}

\item{input}{session input}

\item{output}{session output}

\item{session}{session}

\item{user_selected_sra_sample_info_tibble}{}
}
\description{
A shiny Module.
}
\keyword{internal}
