% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_gene_expression_matrix.R
\name{download_gene_expression_matrix}
\alias{download_gene_expression_matrix}
\alias{download_gene_expression_matrix_server}
\title{mod_download_gene_expression_matrix_ui and mod_download_gene_expression_matrix_server}
\usage{
download_gene_expression_matrix(id)

download_gene_expression_matrix_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{session input}

\item{output}{session output}

\item{session}{session}
}
\description{
A shiny Module.
}
\keyword{internal}
