% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_download_go_data.R
\name{download_go_data_ui}
\alias{download_go_data_ui}
\alias{download_go_data_server}
\title{download_go_data_ui and download_go_data_server}
\usage{
download_go_data_ui(id)

download_go_data_server(input, output, session, ah_data)
}
\arguments{
\item{id}{shiny id}

\item{input}{session input}

\item{output}{session output}

\item{session}{session}

\item{ah_data}{}
}
\description{
A shiny Module.
}
\keyword{internal}
