% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export_base_graphics.R
\name{export_base_graphics_ui}
\alias{export_base_graphics_ui}
\alias{export_base_graphics}
\title{export_base_graphics_ui and export_base_graphics}
\usage{
export_base_graphics_ui(id)

export_base_graphics(
  input,
  output,
  session,
  file_name,
  plot,
  legend_pos = reactive({
     "top"
 }),
  isComplexHeatmap = TRUE
)
}
\arguments{
\item{id}{shiny id}

\item{input}{session input}

\item{output}{session output}

\item{session}{session}

\item{file_name}{internal}

\item{plot}{internal}

\item{legend_pos}{internal}

\item{isComplexHeatmap}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
