% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export_plot.R
\name{export_plot_ui}
\alias{export_plot_ui}
\alias{export_plot}
\title{export_plot_ui export_plot}
\usage{
export_plot_ui(id)

export_plot(input, output, session, file_name, plot)
}
\arguments{
\item{id}{shiny id}

\item{input}{session input}

\item{output}{session output}

\item{session}{session}

\item{file_name}{internal}

\item{plot}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
