% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_functional_analysis.R
\name{functional_analysis_ui}
\alias{functional_analysis_ui}
\alias{functional_analysis_server}
\title{functional_analysis_ui and functional_analysis_server}
\usage{
functional_analysis_ui(id)

functional_analysis_server(
  input,
  output,
  session,
  gene_set = NULL,
  genome = NULL,
  ui_id = NULL
)
}
\arguments{
\item{id}{shiny id}

\item{input}{session input}

\item{output}{session output}

\item{session}{session}

\item{gene_set}{internal}

\item{genome}{internal}

\item{ui_id}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
