% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gene_group_selection.R
\name{gene_group_selection_ui}
\alias{gene_group_selection_ui}
\alias{gene_group_selection}
\title{gene_group_selection_ui and gene_group_selection}
\usage{
gene_group_selection_ui(id)

gene_group_selection(input, output, session, gene_group_info,
  current_session_data_matrix, generate_plot_action)
}
\arguments{
\item{id}{shiny id}

\item{input}{session input}

\item{output}{session input}

\item{session}{session}

\item{gene_group_info}{internal}

\item{current_session_data_matrix}{internal}

\item{generate_plot_action}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
