% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ggplot_fill_and_facet.R
\name{ggplot_fill_and_facet_ui}
\alias{ggplot_fill_and_facet_ui}
\alias{ggplot_fill_and_facet}
\title{ggplot_fill_and_facet_ui and ggplot_fill_and_facet}
\usage{
ggplot_fill_and_facet_ui(id)

ggplot_fill_and_facet(input, output, session, ggplot,
  allow_x_var_selection = TRUE)
}
\arguments{
\item{id}{shiny id}

\item{input}{session input}

\item{output}{session output}

\item{session}{session}

\item{ggplot}{internal}

\item{allow_x_var_selection}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
