% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ggplot_fill_colour.R
\name{ggplot_fill_colour_ui}
\alias{ggplot_fill_colour_ui}
\alias{ggplot_fill_colour}
\title{ggplot_fill_colour_ui and ggplot_fill_colour}
\usage{
ggplot_fill_colour_ui(id, fill_attribute = "Samples")

ggplot_fill_colour(input, output, session, gp, times)
}
\arguments{
\item{id}{shiny id}

\item{fill_attribute}{}

\item{input}{session input}

\item{output}{session output}

\item{session}{session}

\item{gp}{internal}

\item{times}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
