% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_joy_plot_fill_and_facet.R
\name{joy_plot_fill_and_facet_ui}
\alias{joy_plot_fill_and_facet_ui}
\alias{joyplot_fill_and_facet}
\title{joy_plot_fill_and_facet_ui and joyplot_fill_and_facet}
\usage{
joy_plot_fill_and_facet_ui(id)

joyplot_fill_and_facet(input, output, session, joyplot)
}
\arguments{
\item{id}{shiny id}

\item{input}{session input}

\item{output}{session output}

\item{session}{session}

\item{joyplot}{a joy plot object}
}
\description{
A shiny Module.
}
\keyword{internal}
