% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_title_and_axis_label.R
\name{plot_title_and_axis_label_ui}
\alias{plot_title_and_axis_label_ui}
\alias{plot_title_and_axis_label_server}
\title{plot_title_and_axis_label_ui and plot_title_and_axis_label_server}
\usage{
plot_title_and_axis_label_ui(id)

plot_title_and_axis_label_server(input, output, session, plot_title = "",
  axis_x_title = "", axis_y_title = "", color_legend_title = NULL,
  fill_legend_title = NULL, my_ggplot, aspect_ratio = 1,
  apply_theme = TRUE, ...)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{plot_title}{internal}

\item{axis_x_title}{internal}

\item{axis_y_title}{internal}

\item{color_legend_title}{internal}

\item{fill_legend_title}{internal}

\item{my_ggplot}{internal}

\item{aspect_ratio}{internal}

\item{apply_theme}{internal}

\item{...}{Other parameters passed to \code{decorate_ggplot} and ultimately \code{ggplot2::theme}}
}
\description{
A shiny Module.
}
\keyword{internal}
