% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_word_cloud.R
\name{word_cloud_ui}
\alias{word_cloud_ui}
\alias{word_cloud_server}
\title{word_cloud_ui and word_cloud_server}
\usage{
word_cloud_ui(id)

word_cloud_server(input, output, session, parent_id, input_text)
}
\arguments{
\item{id}{shiny id}

\item{input}{session input}

\item{output}{session output}

\item{session}{session}

\item{parent_id}{internal}

\item{input_text}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
