% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{my_filter}
\alias{my_filter}
\title{create column filters}
\usage{
my_filter(col, op, value, env = parent.frame())
}
\arguments{
\item{col}{internal}

\item{op}{internal}

\item{value}{internal}

\item{env}{internal}
}
\value{
internal
}
\description{
create column filters
}
\references{
https://github.com/r-lib/rlang/issues/116
}
\keyword{internal}
