% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{perform_go_enrichmet}
\alias{perform_go_enrichmet}
\title{Perform GO enrichment using clusterProfiler}
\usage{
perform_go_enrichmet(
  genome,
  query_genes,
  pval_cutoff = 1,
  qval_cutoff = 1,
  p_adjust_method = "BH",
  ontology = "Biological Process",
  min_gs_size = 10,
  max_gs_size = 500,
  cross_ref_go_db = TRUE
)
}
\arguments{
\item{genome}{name of the target species genome. It must match to the genome column of \code{ah_data_summary2} object}

\item{query_genes}{internal}

\item{pval_cutoff}{internal}

\item{qval_cutoff}{internal}

\item{p_adjust_method}{internal}

\item{ontology}{internal}

\item{min_gs_size}{internal}

\item{max_gs_size}{internal}

\item{cross_ref_go_db}{internal}
}
\value{
internal
}
\description{
It uses one of the two functions \code{clusterProfiler::enricher} or \code{clusterProfiler::compareCluster}
}
\keyword{internal}
