% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fpkm_files_from_excel_data_functions.R
\name{prepare_annotations_for_fungiexpres_data_matrix}
\alias{prepare_annotations_for_fungiexpres_data_matrix}
\title{Prepare SRA sample annotations required for FungiExpresZ.}
\usage{
prepare_annotations_for_fungiexpres_data_matrix(
  data_matrix,
  fungiexpresz_species,
  fungiexpresz_ref_annotation,
  fungal_sra_run_info_object_path,
  path_sra_db_sqlite_file
)
}
\arguments{
\item{data_matrix}{a FPKM matrix for which SRA annotations to be obtained. Column names of will be considered as SRA id.}

\item{fungiexpresz_species}{a character string denoting a species name for the `data_matrix`. Species must present in the FungiExpresZ GO annotations.}

\item{fungiexpresz_ref_annotation}{a character denoting a reference annotation for the species. Reference annotation must present in the FungiExpresZ GO annotations.}

\item{fungal_sra_run_info_object_path}{a character string denoting path to the `fungal_sra_run_info_table.rds` object.}

\item{path_sra_db_sqlite_file}{a character string denoting path to the `SRAmetadb.sqlite` object.}
}
\value{
a tibble of SRA annotation which can be append to existing `sample_info` object of FungiExpresZ.
}
\description{
Prepare SRA sample annotations required for FungiExpresZ.
}
\keyword{internal}
