% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{reorder_clusters}
\alias{reorder_clusters}
\title{reorder_clusters}
\usage{
reorder_clusters(mat, row_clusters)
}
\arguments{
\item{mat}{: numeric matrix. nrow(mat) must be equal to the length of row_clusters and ncol(mat) must be equal to the column_clusters.}

\item{row_clusters}{: character vector, where each element indicates the row cluster name of mat}
}
\value{
: list
}
\description{
reorder_clusters
}
\keyword{internal}
