% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{shinyInput}
\alias{shinyInput}
\title{create shiny inputs}
\usage{
shinyInput(FUN, len, id, ns, ...)
}
\arguments{
\item{FUN}{A shiny UI function for which vector of defined length need to be created}

\item{len}{An integer denoting length of return vector}

\item{id}{A character string denoting id of UI element}

\item{ns}{An object of class NS}

\item{...}{Other parameters to be pass to argument \code{FUN}}
}
\value{
A character vector of elements \code{FUN} of length \code{len}.
}
\description{
\code{shinyInput} function creates a character vector of length \code{len} for a given shiny input under the argument \code{FUN}
}
\references{
\link{https://stackoverflow.com/questions/45739303/r-shiny-handle-action-buttons-in-data-table}
}
\keyword{internal}
