% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{text_to_clean_document}
\alias{text_to_clean_document}
\title{Convert text into clean document using several functions from the package \code{tm}}
\usage{
text_to_clean_document(
  text,
  remove_numbers = TRUE,
  remove_stop_words = TRUE,
  remove_specific_words = TRUE,
  remove_punctuation = TRUE,
  text_stemming = FALSE,
  specific_words = NULL
)
}
\arguments{
\item{text}{internal}

\item{remove_numbers}{internal}

\item{remove_stop_words}{internal}

\item{remove_specific_words}{internal}

\item{remove_punctuation}{internal}

\item{text_stemming}{internal}

\item{specific_words}{internal}
}
\value{
internal
}
\description{
Convert text into clean document using several functions from the package \code{tm}
}
\keyword{internal}
