% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{text_to_tibble}
\alias{text_to_tibble}
\title{Given a string separated by specified deliminator, the function returns a tibble.}
\usage{
text_to_tibble(text, sep = "\\t")
}
\arguments{
\item{text}{internal}

\item{sep}{internal}
}
\value{
internal
}
\description{
Given a string separated by specified deliminator, the function returns a tibble.
}
\keyword{internal}
