% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{tibble_to_row_clusters}
\alias{tibble_to_row_clusters}
\title{Given a tibble of gene expression matrix, perform row wise clustering}
\usage{
tibble_to_row_clusters(
  x,
  row_ident = 1,
  cols_to_use,
  use_z_score_for_clustering = TRUE,
  output_cols = "zscore",
  num_of_top_var_rows = -1,
  nclust = 4
)
}
\arguments{
\item{x}{internal}

\item{row_ident}{internal}

\item{cols_to_use}{internal}

\item{use_z_score_for_clustering}{internal}

\item{output_cols}{internal}

\item{num_of_top_var_rows}{internal}

\item{nclust}{internal}
}
\value{
internal
}
\description{
Given a tibble of gene expression matrix, perform row wise clustering
}
\keyword{internal}
