% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{tibble_to_rowwise_sd_zs}
\alias{tibble_to_rowwise_sd_zs}
\title{given the tibble and column indices add columns of  row wise zscore and row wise standard deviation}
\usage{
tibble_to_rowwise_sd_zs(
  x,
  cols_to_use,
  std_dev_colname = "sd",
  zscore_colname_suffix = "_zscore"
)
}
\arguments{
\item{x}{}

\item{cols_to_use}{}

\item{std_dev_colname}{}

\item{zscore_colname_suffix}{}
}
\value{
internal
}
\description{
given the tibble and column indices add columns of  row wise zscore and row wise standard deviation
}
\keyword{internal}
